Feat: 책 왼쪽면
